requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
certifi==2020.12.5
    # via requests
chardet==4.0.0
    # via requests
idna==2.10
    # via requests
python-dotenv==0.15.0
    # via -r requirements.in
requests==2.25.1
    # via spotipy
six==1.15.0
    # via spotipy
spotipy==2.16.1
    # via -r requirements.in
urllib3==1.26.2
    # via requests
